# Conflicts:
#	source/core/slang-type-text-util.cpp

+++ source/core/unix/slang-unix-process.cpp
@@ -345,7 +345,7 @@ SlangResult UnixPipeStream::write(const void* buffer, size_t length)
 
     const ssize_t writeResult = ::write(m_fd, buffer, length);
 
-    if (writeResult < 0 || writeResult != length)
+    if (writeResult < 0 || size_t(writeResult) != length)
     {
         return SLANG_FAIL;
     }
@@ -389,7 +389,7 @@ static int pipeCLOEXEC(int pipefd[2])
 #endif
 }
 
-/* static */SlangResult Process::create(const CommandLine& commandLine, Process::Flags flags, RefPtr<Process>& outProcess)
+/* static */SlangResult Process::create(const CommandLine& commandLine, Process::Flags, RefPtr<Process>& outProcess)
 {
     const char* whatFailed = nullptr;
     pid_t childPid;
